Build Tools: Changelog: Merge scripts to single file
